requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
drf-yasg==1.21.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9 